refactor(common): extract useMatchMedia into hooks folder
